lex: support reading literals with underscores.
This commit introduces support for reading/lexing numeric LPC literals
that contain semantically ignored underscores for enhanced readability.

Writing a numeric literal like "int x = 10000000;" makes it hard for
a human being to immediately recognize the order of magnitude of the
number. As of this commit the same literal can be written "int
x = 10_000_000;" which is much faster to recognize as ten million. This
same feature is also made available in the binary, octal and hex literal
notations ("0b1101_1001", "0o777_777", "0xFF_FF").

This brings a popular feature from many other languages (Python, C#,
Java, Perl, Ruby, Rust, etc) to LPC. Similar to the Python PEP-0515[0]
implementation we do not allow leading underscores, or more than one
underscore. E.g. "_1000" or "1__00" are invalid syntax.

[0]: https://peps.python.org/pep-0515/