Don't use transfer_svalue() for apply_return_value
Several functions use apply_return_value to store the result of a callback
call. transfer_svalue() should override any previous value in it, but if
it contains an lvalue, the new value will be assigned to the lvalue instead.
Use pop_apply_value() instead, it makes sure, this doesn't happen.